[WIP] debug

+++ ft_bzero.c
@@ -6,12 +6,12 @@
 /*   By: junghopa <juhnhopa@student.42seoul.>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/12 15:13:07 by junghopa          #+#    #+#             */
-/*   Updated: 2023/12/04 17:16:26 by juggorr          ###   ########.fr       */
+/*   Updated: 2023/12/05 19:11:55 by juggorr          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
-extern void	*ft_memset(void *b, int c, unsigned long len);
+extern void	*ft_memset(void *b, int c, unsigned int len);
 
-void	ft_bzero(void *s, unsigned long n)
+void	ft_bzero(void *s, unsigned int n)
 {
 	ft_memset(s, 0, n);
 }
