+++ ft_memcmp.c
@@ -6,12 +6,12 @@
 /*   By: juggorr <juggorr@gmail.com>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/22 13:19:55 by juggorr           #+#    #+#             */
-/*   Updated: 2023/12/05 15:27:12 by juggorr          ###   ########.fr       */
+/*   Updated: 2023/12/05 19:11:55 by juggorr          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
-int	ft_memcmp(const void *s1, const void *s2, unsigned long n)
+int	ft_memcmp(const void *s1, const void *s2, unsigned int n)
 {
-	unsigned long	idx;
+	unsigned int	idx;
 	unsigned char	*tmp1;
 	unsigned char	*tmp2;
 	unsigned char	res;
