+++ ft_strlen.c
@@ -6,12 +6,12 @@
 /*   By: junghopa <juhnhopa@student.42seoul.>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/09 15:01:10 by junghopa          #+#    #+#             */
-/*   Updated: 2023/11/23 16:19:29 by juggorr          ###   ########.fr       */
+/*   Updated: 2023/12/05 19:15:46 by juggorr          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
-unsigned long	ft_strlen(const char *s)
+unsigned int	ft_strlen(const char *s)
 {
-	unsigned long	cnt;
+	unsigned int	cnt;
 
 	cnt = 0;
 	while (*(s + cnt))
